# docs/api-doc.yml
info:
  title: ''
  version: 0.0.1
  description: ''
host: 'localhost:3000'
schemes:
  - http
basePath: /v1
x-qlik-visibility: private
x-qlik-stability: experimental
swagger: '2.0'
paths:
  /health:
    get:
      description: Returns health status of the service
      produces:
        - application/json
      responses:
        '200':
          description: successful operation
          schema:
            type: string
  '/connections/{id}/authentication':
    get:
      description: Get authentication for a connection
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          type: string
          required: true
      responses:
        '200':
          description: Successful operation
        '404':
          description: Connector not found
  '/connections/{id}':
    get:
      description: Get data from a connector
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          type: string
          required: true
      responses:
        '200':
          description: Successful operation
        '401':
          description: Authentication needed
        '404':
          description: Connector not found
    delete:
      description: Delete a connection
      parameters:
        - name: id
          in: path
          type: string
          required: true
      responses:
        '200':
          description: Successful operation
        '404':
          description: Connector not found
  /connections:
    post:
      description: Create a new connection
      produces:
        - application/json
      parameters:
        - name: body
          in: body
          schema:
            $ref: '#/definitions/connection'
      responses:
        '200':
          description: Successful operation
        '404':
          description: Connector not found
definitions:
  connection:
    type: object
    properties:
      connector:
        description: Name of connector
        type: string
      params:
        description: Settings to pass to the connector
        type: object
responses: {}
parameters: {}
securityDefinitions: {}
tags: []